[UI] Исправлена загрузка страницы настроек Resolve #319
